+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## 0.9.3
+
 - Added a vault scan in progress indicator to empty task lists.
 - Fixed new custom keywords not getting styled in the editor view until restart.
 - Fixed embedded task list not updating with task content changes.